pyproject.toml
[project]
name = "benchmark-visualizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic==1.16.4",
    "anyio==4.10.0",
    "black==22.12.0",
    "blinker==1.9.0",
    "certifi==2025.8.3",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "cloudpickle==2.2.1",
    "contourpy==1.3.2",
    "cycler==0.12.1",
    "databricks-cli==0.18.0",
    "dnspython==2.7.0",
    "docker==6.1.3",
    "email-validator==2.2.0",
    "entrypoints==0.4",
    "exceptiongroup==1.3.0",
    "fastapi==0.116.1",
    "fastapi-cli==0.0.8",
    "fastapi-cloud-cli==0.1.5",
    "flask==2.3.3",
    "fonttools==4.59.0",
    "gitdb==4.0.12",
    "gitpython==3.1.45",
    "greenlet==3.2.3",
    "gunicorn==20.1.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "idna==3.10",
    "igraph==0.10.3",
    "importlib-metadata==5.2.0",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "jmetalpy==1.5.5",
    "joblib==1.5.1",
    "kaleido>=1.2.0",
    "kiwisolver==1.4.8",
    "llvmlite==0.43.0",
    "mako==1.3.10",
    "markdown==3.8.2",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "matplotlib==3.6.2",
    "mdurl==0.1.2",
    "mlflow==2.1.1",
    "mypy-extensions==1.1.0",
    "narwhals==2.0.1",
    "networkx==2.8.8",
    "numba==0.60.0",
    "numpy==1.23.5",
    "oauthlib==3.3.1",
    "packaging==22.0",
    "pandas==1.5.3",
    "passlib==1.7.4",
    "pathspec==0.12.1",
    "patsy==1.0.1",
    "pillow==11.3.0",
    "platformdirs==4.3.8",
    "plotly==6.2.0",
    "protobuf==4.25.8",
    "pyarrow==10.0.1",
    "pydantic==1.10.4",
    "pygments==2.19.2",
    "pyjwt==2.10.1",
    "pyparsing==3.2.3",
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "pytz==2022.7.1",
    "pyyaml==6.0.2",
    "querystring-parser==1.2.4",
    "requests==2.32.4",
    "rich==14.1.0",
    "rich-toolkit==0.14.9",
    "rignore==0.6.4",
    "ruff>=0.14.14",
    "scikit-learn==1.7.1",
    "scipy==1.15.3",
    "sentry-sdk==2.34.1",
    "shap==0.48.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "slicer==0.0.8",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "sqlalchemy==1.4.46",
    "sqlparse==0.5.3",
    "starlette==0.47.2",
    "statsmodels==0.14.5",
    "tabulate==0.9.0",
    "texttable==1.7.0",
    "threadpoolctl==3.6.0",
    "tomli==2.2.1",
    "tqdm==4.66.4",
    "typer==0.16.0",
    "typing-extensions==4.14.1",
    "urllib3==1.26.13",
    "uvicorn==0.35.0",
    "uvloop==0.21.0",
    "watchfiles==1.1.0",
    "websocket-client==1.8.0",
    "websockets==15.0.1",
    "werkzeug==3.1.3",
    "zipp==3.23.0",
]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 128
indent-width = 4

# Assume Python 3.14
target-version = "py314"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`) codes by default.
select = ["E4", "E7", "E9", "F", "E501", "I", "W"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"


[tool.ruff.format]
docstring-code-format = true
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.pyright]
include = [".", "test"]
typeCheckingMode = "standard"
reportUnknownMemberType = "none"
reportUnknownVariableType = "none"
reportUnknownArgumentType = "none"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["test"]

[dependency-groups]
dev = [
    "pyright>=1.1.407",
]
